+++ agents/cap_group_agents/VPC_network/VPC_network_planner.py
@@ -1,0 +1,56 @@
+from agency_swarm import Agent
+from agents.cap_group_agents.planner_instruction import planner_instruction
+from agents.tools.read_json_file.ReadJsonFile import ReadJsonFile
+
+_name = "VPC_network_planner"
+
+_description = """
+负责VPC网络管理能力群的步骤规划
+"""
+
+_group_name = "VPC网络管理能力群"
+
+_input_format = """
+{
+    "title": <任务名称>,
+    "description": <任务描述>,
+}
+"""
+
+_agents = """
+1. **Vpc Agent**: 负责增删改查VPC；
+2. ****: 负责增删改查子网；
+"""
+
+_output_format = """
+{
+    "step_1": {
+        "title": 步骤名称,
+        "id": 步骤ID, 
+        "agent": [agent_name_1, ...],
+        "description": 步骤描述, 
+        "dep": <前置步骤ID列表>,
+    },
+    ...
+}
+"""
+
+_instruction = planner_instruction(_group_name, _input_format, _agents, _output_format)
+
+_tools = [ReadJsonFile]
+
+_file_folder = ""
+
+def create_agent(*, 
+                 description=_description, 
+                 instuction=_instruction, 
+                 tools=_tools, 
+                 files_folder=_file_folder):
+    return Agent(name=_name,
+                 tools=tools,
+                 description=description,
+                 instructions=instuction,
+                 files_folder=_file_folder,
+                 temperature=0.5,
+                 response_format={"type": "json_object"},
+                 max_prompt_tokens=25000,)
